requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2

# Time series analysis
statsmodels>=0.13.0

# Machine learning
xgboost>=1.4.2
lightgbm>=3.3.2

# Visualization
matplotlib>=3.4.2
seaborn>=0.11.1

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Utils
pyyaml>=5.4.1
python-dotenv>=0.19.0